requirements.txt
# Pinned dependencies for security and reproducibility
# Last updated: 2025-11-14
# Use 'pip install --upgrade' to check for newer versions

# HTTP library - latest stable version as of 2025
# https://github.com/psf/requests/releases/tag/v2.32.5
requests==2.32.5

# GitHub API library - latest stable version as of 2025
# https://github.com/PyGithub/PyGithub/releases/tag/v2.8.1
PyGithub==2.8.1

# Testing framework
pytest==8.3.4

# JWT library for GitHub App token generation
# https://github.com/jpadilla/pyjwt/releases
PyJWT==2.9.0
cryptography==43.0.3

